[improve] resolve conflict

+++ manager/src/main/java/org/apache/hertzbeat/manager/controller/MonitorsController.java
@@ -83,7 +83,9 @@ public class MonitorsController {
     public ResponseEntity<Message<Void>> deleteMonitors(
             @Parameter(description = "Monitoring ID List", example = "6565463543") @RequestParam(required = false) List<Long> ids
     ) {
-        monitorService.deleteMonitors(new HashSet<>(ids));
+        if (ids != null && !ids.isEmpty()) {
+            monitorService.deleteMonitors(new HashSet<>(ids));
+        }
         Message<Void> message = Message.success();
         return ResponseEntity.ok(message);
     }
@@ -94,7 +96,9 @@ public class MonitorsController {
     public ResponseEntity<Message<Void>> cancelManageMonitors(
             @Parameter(description = "Monitoring ID List", example = "6565463543") @RequestParam(required = false) List<Long> ids
     ) {
-        monitorService.cancelManageMonitors(new HashSet<>(ids));
+        if (ids != null && !ids.isEmpty()) {
+            monitorService.cancelManageMonitors(new HashSet<>(ids));
+        }
         Message<Void> message = Message.success();
         return ResponseEntity.ok(message);
     }
